К̴̧̗̳̥̙̗̱̜͖̮͉͂̃̊О̶̨̡̧̧͈͕͕̲̻̪͔͔̥̫͕̜͂̓͒̍̀̽͆͆̃͂̏͌ͅМ̶̮̲̜̼̮͆́̓̆̐͘̕͝М̵̣̩̋͗̑̓̂͒͊̀̓ͅͅЙ̷̡̢̧̧̭͕̫̪̻̪͓̘̱̼͔͕̩͛͐̊̒͑̒̒̏̀̔̃͝͝Т̸̳͉̪̖̖͙̻̫̤̬̐͜

+++ src/practice78/EmployeePosition.java
@@ -2,5 +2,5 @@ package practice78;
 
 public interface EmployeePosition {
     String getJobTitle();
-    double calcSalary(double baseSalary);
+    int calcSalary(double baseSalary);
 }
